Conflicts:
	app/assets/javascripts/formBuilder.js
	app/assets/stylesheets/forms.css.scss
	app/models/input.rb
	app/views/forms/_form_editor.html.erb
	app/views/forms/_tool_palette.html.erb
	app/views/forms/show.xml.erb

+++ app/assets/javascripts/formBuilder.js
@@ -1,5 +1,3 @@
-$(document).ready(function(){
-});
 
 var PT = PT || {};
 
@@ -111,6 +109,7 @@ PT.FormModel = function(){
   self.inputs = ko.observableArray([]);
 
   self.addInput = function(event){
+    event.stopPropagation();
     var type = $(event.target).attr("rel");
     var input = new PT.Input();
     input.inEdit(true);
@@ -121,16 +120,17 @@ PT.FormModel = function(){
 
   self.removeInput = function(){
     self.inputs.remove(this);
-
-    $.ajax({
-      url: "/forms/" + PT.form.id + "/inputs/" + this.id,
-      type: "DELETE",
-      contentType: "application/json",
-      dataType: "json"
-    })
-    .done(function(response){
-      console.log(response);
-    });  
+    if(this.id !== ""){
+      $.ajax({
+        url: "/forms/" + PT.form.id + "/inputs/" + this.id,
+        type: "DELETE",
+        contentType: "application/json",
+        dataType: "json"
+      })
+      .done(function(response){
+        console.log(response);
+      });  
+    }
   };
 
   /// Update order of all inputs
@@ -186,7 +186,7 @@ PT.getForm = function(url){
     PT.form.populateInputs(response.inputs);
     ko.applyBindings(PT.form);
 
-    $(document).on("click", "#toolPalette li", PT.form.addInput);
+    $("#toolPalette .toolButton").on("click", PT.form.addInput);
   });
 };
